normalize: fixed substrings
The problem with the previous version is that there are some cases
where one allele is a substring of the other allele (substring at
the start or end) and the other side (end or start, respectively)
has some partial match as well.

This problem was not handled generally, but only for insertions like
GT->GGT, and some cases like GTT->GT failed. The way to detect this
problem was to see if one allele was a substring, and then skip the
leading mismatch test, as it is a substring and all the reference
"should" be context.

The solution is to do always the mismatch test, and doing it with the
shortest allele, to stop comparing as soon as possible. When the c++14's
std::mistmatch is available, those will no longer be special cases,
because the new overload takes both ends to stop in the sortest one.