remove docker: README, Dockerfile, entrypoint.
At the moment the docker is not working (and it's likely that hasn't
been working for a while), so a specific task will be created just
to restore this.

The reasons why it's not working are likely the last dependencies added
(curl, c-ares) and how they don't fit with the install_dependencies.sh script
and the latest CMakeLists.txt.